@@ -1,9 +1,11 @@
-var express = require('express');
-var router = express.Router();
+import express from "express";
+import asyncHandler from "express-async-handler";
+
+const router = express.Router();
 
 /* GET users listing. */
-router.get('/', function(req, res, next) {
-  res.send('respond with a resource');
+router.get("/", function (req, res, next) {
+  res.send("respond with a resource");
 });
 
-module.exports = router;
+export default router;
